chore: fix lint error
